# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/pos_invoices:
    get:
      tags:
        - PosInvoices
      summary: List PosInvoices
      description: Get a paginated list of PosInvoices with optional filtering and sorting
      operationId: listPosInvoices
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosInvoiceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosInvoices
      summary: Create PosInvoice
      description: Create a new PosInvoice
      operationId: createPosInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosInvoiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosInvoice'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/pos_invoice_items:
    get:
      tags:
        - PosInvoiceItems
      summary: List PosInvoiceItems
      description: Get a paginated list of PosInvoiceItems with optional filtering and sorting
      operationId: listPosInvoiceItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosInvoiceItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PosInvoiceItems
      summary: Create PosInvoiceItem
      description: Create a new PosInvoiceItem
      operationId: createPosInvoiceItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePosInvoiceItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PosInvoiceItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PosInvoice:
      type: object
      required:
        - id
        - company_id
        - pos_profile_id
        - opening_entry_id
        - receipt_number
        - posting_at
        - net_total
        - tax_total
        - grand_total
        - rounding_adjustment
        - rounded_total
        - paid_total
        - change_due
        - is_return
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        receipt_number:
          type: string
          maxLength: 40
        posting_at:
          type: string
          format: date-time
        net_total:
          type: string
        tax_total:
          type: string
        grand_total:
          type: string
        rounding_adjustment:
          type: string
        rounded_total:
          type: string
        paid_total:
          type: string
        change_due:
          type: string
        billing_invoice_id:
          nullable: true
          type: string
          format: uuid
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        is_return:
          type: boolean
        return_against:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    CreatePosInvoiceRequest:
      type: object
      required:
        - company_id
        - pos_profile_id
        - opening_entry_id
        - receipt_number
        - posting_at
        - net_total
        - tax_total
        - grand_total
        - rounding_adjustment
        - rounded_total
        - paid_total
        - change_due
        - is_return
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_profile_id:
          type: string
          format: uuid
        opening_entry_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          nullable: true
          type: string
          format: uuid
        receipt_number:
          type: string
          maxLength: 40
        posting_at:
          type: string
          format: date-time
        net_total:
          type: string
        tax_total:
          type: string
        grand_total:
          type: string
        rounding_adjustment:
          type: string
        rounded_total:
          type: string
        paid_total:
          type: string
        change_due:
          type: string
        billing_invoice_id:
          nullable: true
          type: string
          format: uuid
        payment_entry_id:
          nullable: true
          type: string
          format: uuid
        is_return:
          type: boolean
        return_against:
          nullable: true
          type: string
          format: uuid
        status:
          type: string
        metadata:
          type: object
...
    PosInvoiceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosInvoice'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PosInvoiceItem:
      type: object
      required:
        - id
        - company_id
        - pos_invoice_id
        - item_id
        - quantity
        - unit_price
        - discount_amount
        - net_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        pos_invoice_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        discount_amount:
          type: string
        net_amount:
          type: string
        revenue_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    CreatePosInvoiceItemRequest:
      type: object
      required:
        - company_id
        - pos_invoice_id
        - item_id
        - quantity
        - unit_price
        - discount_amount
        - net_amount
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        pos_invoice_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        discount_amount:
          type: string
        net_amount:
          type: string
        revenue_account_id:
          nullable: true
          type: string
          format: uuid
        metadata:
          type: object
...
    PosInvoiceItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PosInvoiceItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []