# src/presentation/http/pos_invoice_item_handler.rs
//! PosInvoiceItem REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PosInvoiceItemService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePosInvoiceItemDto, UpdatePosInvoiceItemDto, PatchPosInvoiceItemDto, PosInvoiceItemResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PosInvoiceItemError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PosInvoiceItemError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PosInvoiceItemError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "POSINVOICEITEM_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "POSINVOICEITEM_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSINVOICEITEM_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "POSINVOICEITEM_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PosInvoiceItem.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /pos_invoice_items | List with pagination |
/// | POST | /pos_invoice_items | Create new |
/// | GET | /pos_invoice_items/:id | Get by ID |
/// | PUT | /pos_invoice_items/:id | Full update |
/// | PATCH | /pos_invoice_items/:id | Partial update |
/// | DELETE | /pos_invoice_items/:id | Soft delete |
/// | POST | /pos_invoice_items/bulk | Bulk create |
/// | POST | /pos_invoice_items/upsert | Upsert |
/// | GET | /pos_invoice_items/trash | List deleted |
/// | POST | /pos_invoice_items/:id/restore | Restore |
/// | DELETE | /pos_invoice_items/empty | Empty trash |
/// | GET | /pos_invoice_items/:id/deleted | Get deleted by ID |
/// | DELETE | /pos_invoice_items/trash/:id | Permanent delete from trash |
/// | GET | /pos_invoice_items/count | Count active entities |
/// | GET | /pos_invoice_items/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PosInvoiceItemService::with_repository(repository));
/// let router = create_pos_invoice_item_routes(service);
/// ```
pub fn create_pos_invoice_item_routes(service: Arc<PosInvoiceItemService>) -> Router {
    BackboneCrudHandler::<PosInvoiceItemService, PosInvoiceItem, CreatePosInvoiceItemDto, UpdatePosInvoiceItemDto, PosInvoiceItemResponseDto>::routes(
        service,
        "/pos_invoice_items",
    )
}

/// Create Axum router with only the read (GET) endpoints for PosInvoiceItem.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_pos_invoice_item_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_pos_invoice_item_read_routes(service: Arc<PosInvoiceItemService>) -> Router {
    BackboneCrudHandler::<PosInvoiceItemService, PosInvoiceItem, CreatePosInvoiceItemDto, UpdatePosInvoiceItemDto, PosInvoiceItemResponseDto>::read_routes(
        service,
        "/pos_invoice_items",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PosInvoiceItem.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_pos_invoice_item_write_routes(service: Arc<PosInvoiceItemService>) -> Router {
    BackboneCrudHandler::<PosInvoiceItemService, PosInvoiceItem, CreatePosInvoiceItemDto, UpdatePosInvoiceItemDto, PosInvoiceItemResponseDto>::write_routes(
        service,
        "/pos_invoice_items",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_pos_invoice_item_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PosInvoiceItemService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_pos_invoice_item_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}